fixed graph import

+++ core/plugins/dependency/node.go
@@ -3,14 +3,14 @@ package dependency
 import "github.com/secure2work/nori/core/plugins/meta"
 
 type Node interface {
-	ID() int
+	ID() int64
 }
 
 type PNode interface {
 	PID() meta.ID
 }
 
-func NewNode(id int, ID meta.ID) Node {
+func NewNode(id int64, ID meta.ID) Node {
 	return node{
 		id:  id,
 		pid: ID,
@@ -18,11 +18,11 @@ func NewNode(id int, ID meta.ID) Node {
 }
 
 type node struct {
-	id  int
+	id  int64
 	pid meta.ID
 }
 
-func (d node) ID() int {
+func (d node) ID() int64 {
 	return d.id
 }
